Added progress window for pinging challonge.com

+++ src/GUI/TemplateWindow.java
@@ -55,14 +55,14 @@ public class TemplateWindow {
     }
 
     public int getCenter(JLabel l) {
-        return offset+((int)(window.getWidth()/2)) - (int)(l.getWidth()/2);
+        return offset+((window.getWidth()/2) - (l.getWidth()/2));
     }
 
     public int getCenter(JButton b) {
-        return offset+((int)(window.getWidth()/2)) - (int)(b.getWidth()/2);
+        return offset+((window.getWidth()/2) - (b.getWidth()/2));
     }
 
     public int getCenter(JTextField f) {
-        return offset+((int)(window.getWidth()/2)) - (int)(f.getWidth()/2);
+        return offset+((window.getWidth()/2) - (f.getWidth()/2));
     }
 }
